requirements.txt
# Dev environment
pip
autopep8

# App
flask
google-generativeai
python-dotenv
Pillow
